Разбил спеки в features по папками
