+++ spec/features/auth/sign_up_spec.rb
@@ -1,4 +1,4 @@
-require_relative 'features_helper'
+require_relative '../features_helper'
 
 feature 'User can register', %q{
   In order to be able ask questions
